refactor: move site config from utils to site-config
- Moved the site configuration like title, description, logo, breadcrumb routes etc from utils/config.ts to a new file utils/site-config.ts.
- Updated imports in multiple files to use the new site-config instead of old config.
