+++ src/layouts/Layout.astro
@@ -5,7 +5,7 @@ import Footer from "#components/astro/Footer.astro";
 import Header from "#components/astro/Header.astro";
 import Sidebar from "#components/astro/Sidebar.astro";
 import AstroBreadcrumb from "#components/react/astro-breadcrumb.tsx";
-import { SITE_DESCRIPTION, SITE_TITLE } from "#utils/config";
+import { SITE_DESCRIPTION, SITE_TITLE } from "#utils/site-config";
 import "@fpkit/react/styles";
 
 // types
@@ -65,4 +65,5 @@ import "../styles/index.css";
     </main>
     <Footer />
   </body>
+  #utils/site-config
 </html>
