+++ src/pages/posts/[page].astro
@@ -1,6 +1,6 @@
 ---
 import { getCollection } from "astro:content";
-import { PAGINATION_COUNT } from "#utils/config";
+import { PAGINATION_COUNT } from "#utils/site-config";
 import Layout from "#layouts/Layout.astro";
 import BlogPosts from "#components/astro/BlogPosts.astro";
 import Pagination from "#components/astro/Pagination.astro";
@@ -46,3 +46,4 @@ const { page } = Astro.props as any;
     </div>
   </nav>
 </Layout>
+#utils/site-config
